Cleaned signed/unsigned types mixtures, got rid from annoying compiler warnings.
git-svn-id: http://svn.osgeo.org/geos/trunk@1986 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/algorithm/HCoordinate.cpp
@@ -194,7 +194,7 @@ HCoordinate::getX() const
 	long double a = x/w;
 
 	// finite() also checks for NaN
-	if ( ! finite(a) )
+	if ( ! finite(static_cast<double>(a)) )
 	{
 		throw  NotRepresentableException();
 	}
@@ -208,7 +208,7 @@ HCoordinate::getY() const
 	long double a = y/w;
 
 	// finite() also checks for NaN
-	if ( ! finite(a) )
+	if ( ! finite(static_cast<double>(a)) )
 	{
 		throw  NotRepresentableException();
 	}
@@ -219,7 +219,7 @@ HCoordinate::getY() const
 void
 HCoordinate::getCoordinate(Coordinate &ret) const
 {
-	ret=Coordinate(getX(), getY());
+	ret = Coordinate(static_cast<double>(getX()), static_cast<double>(getY()));
 }
 
 std::ostream& operator<< (std::ostream& o, const HCoordinate& c)
